cast magma_int_t up to 64-bits using (long long); in testing
git-svn-id: https://icl.cs.utk.edu/svn/magma/trunk@9389 02bb5f2a-90ca-11de-bb24-8aedcc835bc8

+++ testing/magma_util.cpp
@@ -416,10 +416,10 @@ void magma_opts::parse_opts( int argc, char** argv )
             // save in environment variable, so magma_num_gpus() picks it up
             char env_num_gpus[20];  // space for "MAGMA_NUM_GPUS=", 4 digits, and nil
             #if defined( _WIN32 ) || defined( _WIN64 )
-                snprintf( env_num_gpus, sizeof(env_num_gpus), "MAGMA_NUM_GPUS=%ld", long(std::abs(this->ngpu)) );
+                snprintf( env_num_gpus, sizeof(env_num_gpus), "MAGMA_NUM_GPUS=%lld", (long long) std::abs(this->ngpu) );
                 putenv( env_num_gpus );
             #else
-                snprintf( env_num_gpus, sizeof(env_num_gpus), "%ld", long(std::abs(this->ngpu)) );
+                snprintf( env_num_gpus, sizeof(env_num_gpus), "%lld", (long long) std::abs(this->ngpu) );
                 setenv( "MAGMA_NUM_GPUS", env_num_gpus, true );
             #endif
         }
