+++ testing/testing_dgeev.cpp
@@ -104,9 +104,9 @@ int main( int argc, char** argv)
     // pass ngpu = -1 to test multi-GPU code using 1 gpu
     magma_int_t abs_ngpu = std::abs( opts.ngpu );
     
-    printf("%% jobvl = %s, jobvr = %s, ngpu = %ld\n",
+    printf("%% jobvl = %s, jobvr = %s, ngpu = %lld\n",
            lapack_vec_const(opts.jobvl), lapack_vec_const(opts.jobvr),
-           long(abs_ngpu) );
+           (long long) abs_ngpu );
     
     printf("%%   N   CPU Time (sec)   GPU Time (sec)   |W_magma - W_lapack| / |W_lapack|\n");
     printf("%%==========================================================================\n");
@@ -158,8 +158,8 @@ int main( int argc, char** argv)
             }
             gpu_time = magma_wtime() - gpu_time;
             if (info != 0) {
-                printf("magma_dgeev returned error %ld: %s.\n",
-                       long(info), magma_strerror( info ));
+                printf("magma_dgeev returned error %lld: %s.\n",
+                       (long long) info, magma_strerror( info ));
             }
 
             /* =====================================================================
@@ -322,8 +322,8 @@ int main( int argc, char** argv)
                                    h_work, lwork, &info );
                 }
                 if (info != 0) {
-                    printf("magma_dgeev (case V, V) returned error %ld: %s.\n",
-                           long(info), magma_strerror( info ));
+                    printf("magma_dgeev (case V, V) returned error %lld: %s.\n",
+                           (long long) info, magma_strerror( info ));
                 }
                 
                 // ----------
@@ -343,8 +343,8 @@ int main( int argc, char** argv)
                 //                   h_work, lwork, &info );
                 //}
                 //if (info != 0) {
-                //    printf("magma_dgeev (case N, N) returned error %ld: %s.\n",
-                //           long(info), magma_strerror( info ));
+                //    printf("magma_dgeev (case N, N) returned error %lld: %s.\n",
+                //           (long long) info, magma_strerror( info ));
                 //}
                 //
                 //// Do test 5: W(full) = W(partial, W only)
@@ -369,8 +369,8 @@ int main( int argc, char** argv)
                                    h_work, lwork, &info );
                 }
                 if (info != 0) {
-                    printf("magma_dgeev (case N, V) returned error %ld: %s.\n",
-                           long(info), magma_strerror( info ));
+                    printf("magma_dgeev (case N, V) returned error %lld: %s.\n",
+                           (long long) info, magma_strerror( info ));
                 }
                 
                 // Do test 6: W(full) = W(partial, W and VR)
@@ -402,8 +402,8 @@ int main( int argc, char** argv)
                                    h_work, lwork, &info );
                 }
                 if (info != 0) {
-                    printf("magma_dgeev (case V, N) returned error %ld: %s.\n",
-                           long(info), magma_strerror( info ));
+                    printf("magma_dgeev (case V, N) returned error %lld: %s.\n",
+                           (long long) info, magma_strerror( info ));
                 }
                 
                 // Do test 7: W(full) = W(partial, W and VL)
@@ -434,8 +434,8 @@ int main( int argc, char** argv)
                                  h_work, &lwork, &info );
                 cpu_time = magma_wtime() - cpu_time;
                 if (info != 0) {
-                    printf("lapackf77_dgeev returned error %ld: %s.\n",
-                           long(info), magma_strerror( info ));
+                    printf("lapackf77_dgeev returned error %lld: %s.\n",
+                           (long long) info, magma_strerror( info ));
                 }
                 
                 // check | W_magma - W_lapack | / | W |
@@ -467,14 +467,14 @@ int main( int argc, char** argv)
                 error  = magma_cblas_dznrm2( N, w1copy, 1 );
                 error /= magma_cblas_dznrm2( N, w2copy, 1 );
                 
-                printf("%5ld   %7.2f          %7.2f          %8.2e   %s\n",
-                       long(N), cpu_time, gpu_time,
+                printf("%5lld   %7.2f          %7.2f          %8.2e   %s\n",
+                       (long long) N, cpu_time, gpu_time,
                        error, (error < tolulp ? "ok" : "failed"));
                 status += ! (error < tolulp);
             }
             else {
-                printf("%5ld     ---            %7.2f\n",
-                       long(N), gpu_time);
+                printf("%5lld     ---            %7.2f\n",
+                       (long long) N, gpu_time);
             }
             if ( opts.check ) {
                 // -1 indicates test was not run
